move tests into one file
